% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnh.app-package.R
\docType{package}
\name{gnh.app-package}
\alias{gnh.app}
\alias{gnh.app-package}
\title{gnh.app: Bhutan National Happiness Index}
\description{
This is a golem app that presents the results of the Bhutan Gross National Happiness Survey.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AyushBipinPatel/bhutan-gnh-app}
  \item Report bugs at \url{https://github.com/AyushBipinPatel/bhutan-gnh-app/issues}
}

}
\author{
\strong{Maintainer}: Ayush Patel \email{ayush.ap58@gmail.com}

}
\keyword{internal}
